+++ src/Ops.DataProvider.SqlServer.Promenade/Migrations/20220119230535_prom_v1.0.0.166.Designer.cs
@@ -10,15 +10,15 @@ using Ocuda.Ops.DataProvider.SqlServer.Promenade;
 namespace Ocuda.Ops.DataProvider.SqlServer.Promenade.Migrations
 {
     [DbContext(typeof(Context))]
-    [Migration("20211116192713_nav1")]
-    partial class nav1
+    [Migration("20220119230535_prom_v1.0.0.166")]
+    partial class prom_v100166
     {
         protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
                 .HasAnnotation("Relational:MaxIdentifierLength", 128)
-                .HasAnnotation("ProductVersion", "5.0.12")
+                .HasAnnotation("ProductVersion", "5.0.13")
                 .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
             modelBuilder.Entity("Microsoft.AspNetCore.DataProtection.EntityFrameworkCore.DataProtectionKey", b =>
@@ -850,6 +850,32 @@ namespace Ocuda.Ops.DataProvider.SqlServer.Promenade.Migrations
                     b.ToTable("LocationHoursOverrides");
                 });
 
+            modelBuilder.Entity("Ocuda.Promenade.Models.Entities.LocationProductMap", b =>
+                {
+                    b.Property<int>("Id")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("int")
+                        .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
+
+                    b.Property<string>("ImportLocation")
+                        .HasMaxLength(255)
+                        .HasColumnType("nvarchar(255)");
+
+                    b.Property<int>("LocationId")
+                        .HasColumnType("int");
+
+                    b.Property<int>("ProductId")
+                        .HasColumnType("int");
+
+                    b.HasKey("Id");
+
+                    b.HasIndex("LocationId");
+
+                    b.HasIndex("ProductId");
+
+                    b.ToTable("LocationProductMaps");
+                });
+
             modelBuilder.Entity("Ocuda.Promenade.Models.Entities.Navigation", b =>
                 {
                     b.Property<int>("Id")
@@ -1271,6 +1297,9 @@ namespace Ocuda.Ops.DataProvider.SqlServer.Promenade.Migrations
                     b.Property<int?>("Season")
                         .HasColumnType("int");
 
+                    b.Property<int?>("ShowNotesSegmentId")
+                        .HasColumnType("int");
+
                     b.Property<string>("Stub")
                         .IsRequired()
                         .HasMaxLength(50)
@@ -1302,6 +1331,9 @@ namespace Ocuda.Ops.DataProvider.SqlServer.Promenade.Migrations
                         .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
+                    b.Property<int>("CacheInventoryMinutes")
+                        .HasColumnType("int");
+
                     b.Property<DateTime>("CreatedAt")
                         .HasColumnType("datetime2");
 
@@ -1311,10 +1343,16 @@ namespace Ocuda.Ops.DataProvider.SqlServer.Promenade.Migrations
                     b.Property<bool>("IsActive")
                         .HasColumnType("bit");
 
+                    b.Property<bool>("IsVisibleToPublic")
+                        .HasColumnType("bit");
+
                     b.Property<string>("Name")
                         .IsRequired()
                         .HasMaxLength(255)
                         .HasColumnType("nvarchar(255)");
+
+                    b.Property<int?>("SegmentId")
+                        .HasColumnType("int");
 
                     b.Property<string>("Slug")
                         .IsRequired()
@@ -1346,7 +1384,16 @@ namespace Ocuda.Ops.DataProvider.SqlServer.Promenade.Migrations
                     b.Property<int>("CreatedBy")
                         .HasColumnType("int");
 
-                    b.Property<int>("InventoryStatus")
+                    b.Property<int?>("ItemCount")
+                        .HasColumnType("int");
+
+                    b.Property<int>("ManyThreshhold")
+                        .HasColumnType("int");
+
+                    b.Property<DateTime?>("ThreshholdUpdatedAt")
+                        .HasColumnType("datetime2");
+
+                    b.Property<int?>("ThreshholdUpdatedBy")
                         .HasColumnType("int");
 
                     b.Property<DateTime?>("UpdatedAt")
@@ -1941,6 +1988,25 @@ namespace Ocuda.Ops.DataProvider.SqlServer.Promenade.Migrations
                         .OnDelete(DeleteBehavior.Restrict);
 
                     b.Navigation("Location");
+                });
+
+            modelBuilder.Entity("Ocuda.Promenade.Models.Entities.LocationProductMap", b =>
+                {
+                    b.HasOne("Ocuda.Promenade.Models.Entities.Location", "Location")
+                        .WithMany()
+                        .HasForeignKey("LocationId")
+                        .OnDelete(DeleteBehavior.Restrict)
+                        .IsRequired();
+
+                    b.HasOne("Ocuda.Promenade.Models.Entities.Product", "Product")
+                        .WithMany()
+                        .HasForeignKey("ProductId")
+                        .OnDelete(DeleteBehavior.Restrict)
+                        .IsRequired();
+
+                    b.Navigation("Location");
+
+                    b.Navigation("Product");
                 });
 
             modelBuilder.Entity("Ocuda.Promenade.Models.Entities.Navigation", b =>