+++ src/Ops.DataProvider.SqlServer.Promenade/Migrations/20220119230535_prom_v1.0.0.166.cs
@@ -2,7 +2,7 @@
 
 namespace Ocuda.Ops.DataProvider.SqlServer.Promenade.Migrations
 {
-    public partial class nav1 : Migration
+    public partial class prom_v100166 : Migration
     {
         protected override void Up(MigrationBuilder migrationBuilder)
         {